AnsibleRole{,Namespace}{Detail,List} - rename for consistency
No-Issue
