+++ src/containers/index.ts
@@ -4,6 +4,10 @@ export { default as AnsibleRemoteList } from './ansible-remote/list';
 export { default as AnsibleRepositoryDetail } from './ansible-repository/detail';
 export { default as AnsibleRepositoryEdit } from './ansible-repository/edit';
 export { default as AnsibleRepositoryList } from './ansible-repository/list';
+export { default as AnsibleRoleNamespaceDetail } from './ansible-role/namespace-detail';
+export { default as AnsibleRoleNamespaceList } from './ansible-role/namespace-list';
+export { default as AnsibleRoleDetail } from './ansible-role/role-detail';
+export { default as AnsibleRoleList } from './ansible-role/role-list';
 export { default as CertificationDashboard } from './certification-dashboard/certification-dashboard';
 export { default as CollectionContent } from './collection-detail/collection-content';
 export { default as CollectionDependencies } from './collection-detail/collection-dependencies';
@@ -22,10 +26,6 @@ export { default as ExecutionEnvironmentRegistryList } from './execution-environ
 export { default as GroupDetail } from './group-management/group-detail';
 export { default as GroupList } from './group-management/group-list';
 export { default as LandingPage } from './landing/landing-page';
-export { default as LegacyNamespace } from './legacy-namespaces/legacy-namespace';
-export { default as LegacyNamespaces } from './legacy-namespaces/legacy-namespaces';
-export { default as LegacyRole } from './legacy-roles/legacy-role';
-export { default as LegacyRoles } from './legacy-roles/legacy-roles';
 export { default as LoginPage } from './login/login';
 export { default as MyImports } from './my-imports/my-imports';
 export { default as NamespaceDetail } from './namespace-detail/namespace-detail';
